Fix a bit of old code docs, that no longer reflect the current state of our knowledge.
Thanks go to CodeWalker, for lending his time and finding that crusty
nugget :)

+++ Components/Common/Bitstream/include/BitStream.h
@@ -28,7 +28,7 @@ Description: The BitStream class allows it's user to manipulate data in
 #define BS_F32         7
 
 #define BITS_PER_BYTE  8
-#define BITS_PER_DWORD (BITS_PER_BYTE * sizeof(uint32_t))
+#define BITS_PER_DWORD (BITS_PER_BYTE * uint32_t(sizeof(uint32_t)))
 
 
 //	Macros
